requirements.txt
django>=5,<6
pandas==3.0.2
psycopg[binary]>=3.1,<4
PyYAML>=6.0,<7